+++ Model/essLinacInc/Linac.h
@@ -63,7 +63,7 @@ class Linac : public attachSystem::ContainedComp,
   double floorThick; ///< Thickness of floor
   double floorWidthLeft; ///< floor width towards x+
   double floorWidthRight; ///< floor width towards x-
-  int nAirLayers; ///< number of layers in the air of the tunnel (along the proton beam)
+  size_t nAirLayers; ///< number of layers in the air of the tunnel (along the proton beam)
 
   int airMat;                    ///< air material
   int wallMat;                   ///< wall material
@@ -76,7 +76,7 @@ class Linac : public attachSystem::ContainedComp,
   std::vector<std::shared_ptr<DTL> > dtl; ///< array of DTL sections
 
   void layerProcess(Simulation& System, const std::string& cellName,
-		    const size_t& lpS, const size_t& lsS, const int&, const int&);
+		    const long int& lpS, const long int& lsS, const size_t&, const int&);
 
   void createDTL(Simulation&, const long int);
   void buildTSW(Simulation& system) const;
