+++ Model/essLinacInc/TSW.h
@@ -54,7 +54,8 @@ class TSW : public attachSystem::ContainedComp,
   int airMat;                   ///< Air material
 
   void layerProcess(Simulation& System, const std::string& cellName,
-		    const size_t& lpS, const size_t& lsS, const size_t&, const int&);
+		    const long int& lpS, const long int& lsS, const size_t&,
+		    const int&);
 
   void populate(const FuncDataBase&);
   void createUnitVector(const attachSystem::FixedComp&,
